- Some random attempts to build upload heap and simplistic texture manager in AZDO fashion. - PNG loading. - Minor fixes

+++ src/graphics/opengl-backend/gl-shader.cc
@@ -113,7 +113,7 @@ GLShaderProgram::GLShaderProgram(GLuint prog, ResourceTableDescription* res_tabl
 
 GLShaderProgram::~GLShaderProgram()
 {
-    for(size_t i = 0; i < m_ResourceTableCount; ++i)
+    for(uint32 i = 0; i < m_ResourceTableCount; ++i)
     {
         DestroyPackedData(m_ResourceTables[i]);
     }
